+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.configuration.AppConfig;
+import racingcar.config.AppConfig;
 import racingcar.controller.RacingController;
 
 public class Application {
